% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.data.train}
\alias{avm.data.train}
\title{data processing for fitting}
\usage{
avm.data.train(tabln.vec = NULL, proptype = "11", modelpath1 = NULL,
  city_code = NULL, var.names.except = c("name", "ha_cl_code", "buildyear",
  "volume_rate", "greening_rate", "proptype", "ha_cl_name", "dist_code",
  "dist_name", "x_", "y_", "bldg_code", "year", "month"))
}
\arguments{
\item{tabln.vec}{basic data list generated by \code{\link{loadData}}}

\item{proptype}{type code of properties.The default values is 11 indicting housing.}

\item{var.names.except}{Variable names which will be exclued in the fitting. See detail}

\item{bldg_code}{type code of building.The default is 11}
}
\value{
trainning data for AVM
}
\description{
This function is useful for generating data for fitting
}
\details{
The default value of var.names.except is \code{c('city_code','proptype','year','month','salecount','rentcount','name','ha_cl_code','ha_cl_name','dist_code','dist_name')}
}
