% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.fit}
\alias{avm.fit}
\title{avm modeling}
\usage{
avm.fit(data.train = NULL, dependentY = "saleprice",
  independentX = "~.-ha_code-saleprice", model.list = list(ols = T, rlm = F,
  svm = F, rft = F, gwr = F, xgb = F), min.missing.rate = 0.4,
  remove.dup = T, ntree = 500, nround = 1000)
}
\arguments{
\item{data.train}{data for fitting generated by \code{\link{avm.data.train}} or \code{\link{data.split}}}

\item{dependentY}{dependent variable. The default is 'saleprice'}

\item{independentX}{independent variables. Using var exception,see detail.}

\item{model.list}{The list of model. The default is \code{list(ols=T,rlm=T,svm=T,rft=T,gwr=T,xgb=T)}}

\item{min.missing.rate}{The rate of missing rate for keeping, The default is 0.4 which indicting when the rate of a feature is more than 40 percent, it will be removed from data}

\item{remove.dup}{Logic variable indicting whether to remove duplicated coordinates.The default is T}

\item{ntree}{numerical value indiciting number of subtrees for random forests.The default is 1000}

\item{nround}{numerical value indicting iteration of xgboost.The default is 1000}
}
\value{
The list of AVM models
}
\description{
model fitting function
}
\details{
parameter \code{independentX} should be written as a formula. For example, '~.-yearmonth-ymid-ha_code' means all variables except yearmonth,ymid and ha_code will be performed as features in the models.

It can also written as '~bldgrea + x + y' indicting bldgrea ,x and y will included in the sets of features . The default value is \code{paste0('~.-yearmonth-ymid-ha_code-volume_rate-greening_rate','-x_-y_-saleprice-rentprice-unitprice','-rentbldgarea-salebldgarea-bldg_type-bldg_code')}
}
