% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{avm.model.restore}
\alias{avm.model.restore}
\title{restore models from files}
\usage{
avm.model.restore(city_code = NULL, yearmonth = "2017-7",
  avm.model.store.path = "./models.avm", proptype = "11",
  dependentY = "saleprice")
}
\arguments{
\item{city_code}{code of city,required.}

\item{yearmonth}{characters indicting year and month. The default value is '2017-7'}

\item{avm.model.store.path}{The path for saving models}

\item{proptype}{type code of properties.The default values is 11 indicting housing.}

\item{dependentY}{dependent variable. The default is 'saleprice'}
}
\value{
model object
}
\description{
This function can read models from files
}
