% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{avm.model.store}
\alias{avm.model.store}
\title{store models as files}
\usage{
avm.model.store(city_code = NULL, avm.models = NULL, yearmonth = "2017-7",
  avm.model.store.path = "./models.avm", proptype = "11",
  dependentY = "saleprice")
}
\arguments{
\item{city_code}{code of city,required.}

\item{avm.models}{object of models,generated by \code{\link{avm.fit}}}

\item{yearmonth}{characters indicting year and month. The default value is '2017-7'}

\item{avm.model.store.path}{The path for saving models}

\item{proptype}{type code of properties.The default values is 11 indicting housing.}

\item{dependentY}{dependent variable. The default is 'saleprice'}
}
\description{
This function can save model as files.
}
