% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.pred.data}
\alias{avm.pred.data}
\title{generate data for prediction}
\usage{
avm.pred.data(train.models = NULL, proptype = "11", x = NULL, y = NULL,
  tabln.vec = NULL, poi.data.list = NULL, spatial = F)
}
\arguments{
\item{train.models}{object of models provided by \code{\link{avm.fit}},required pameter.}

\item{proptype}{type code of properties.The default is 11 which means housing.}

\item{x}{lon,required}

\item{y}{lat,required}

\item{tabln.vec}{basic data provided by \code{\link{loadData}}}

\item{poi.data.list}{data of POI ,required . It will be provided by \code{\link{getpoi.ras}}}

\item{spatial}{logic parameter indicintg whether to return spatial variables. The default is F}

\item{bldg_code}{tyep code of building.The default is 11}

\item{bldg_area}{area of building . The default value is 80}

\item{br}{number beedroom. The default is 2}

\item{buildyear}{year of building. The default is current year.}
}
\value{
data used for predictiong
}
\description{
This function will generate data for given data
}
