% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.pred}
\alias{avm.pred}
\title{avm prediction function}
\usage{
avm.pred(train.models = NULL, model.list = list(ols = T, rlm = T, svm = T,
  rft = T, gwr = T, xgb = T), newdata = NULL, testing = F, remove.dup = T)
}
\arguments{
\item{train.models}{obejct of models provided by \code{\link{avm.fit}}}

\item{model.list}{The list of model .The default is \code{list(ols=T,rlm=T,svm=T,rft=T,gwr=T,xgb=T)}}

\item{newdata}{Testing data or unkown data for prediction. Tesing data is provided by \code{\link{data.split}}}

\item{testing}{Logic value indicting whether to test.The default is F}

\item{remove.dup}{Logic variable indicting whether to remove duplicated coordinates.The default is T}
}
\value{
prediction value . It will contain rmse and mae for testing if testing is T
}
\description{
This function will predict according given data
}
