% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{avm.predall}
\alias{avm.predall}
\title{This function is used for predicating in all spatial area}
\usage{
avm.predall(train.models, tabln.vec, poi.data.list, data.sets.all, model.list,
  bldg_code, remove.dup = T)
}
\arguments{
\item{train.models}{avmmodels made before}

\item{tabln.vec}{containing community basal information}

\item{poi.data.list}{containg poi information}

\item{data.sets.all}{all valuable data}

\item{model.list}{select your models}
}
\value{
trainning data for AVM
}
\description{
This function is used for predicating in all spatial area
}
