% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{data.split}
\alias{data.split}
\title{data spliting}
\usage{
data.split(data, train.rate = 0.7)
}
\arguments{
\item{data}{all data sets provided by \code{\link{avm.data.train}}}

\item{train.rate}{the rate for training. Default value is 0.7, which means 70\% of data will be used in the fitting}
}
\value{
a list contain train and test data
}
\description{
This function is useful for spliting training and testing data
}
