% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dbconnect}
\alias{dbconnect}
\title{DB connection for testing}
\usage{
dbconnect(user = "avm", password = "gwr", host = "db", port = 3306,
  dbname = "cityre")
}
\arguments{
\item{user}{username for connection, default value is avm}

\item{password}{password for connection, default value is gwr}

\item{host}{host of db server}

\item{port}{port number of db server}

\item{dbname}{dbname of db server}
}
\description{
connection for testing. It will print "It Works!" if success.
}
\examples{
con<-dbconnect()
dbListTable(conn = con)
dbDisconnect(con)
}
