% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dbconnect2}
\alias{dbconnect2}
\title{dbconnect2}
\usage{
dbconnect2(user = "wangcl", password = "wangchenliang_CR_2017",
  host = "10.11.10.34", dbname = "cityre_beijing", port = 3307)
}
\arguments{
\item{user}{usename}

\item{password}{password}

\item{host}{host}

\item{dbname}{dbname}

\item{port}{port}
}
\description{
connection for testing. It will print "It Works!" if success.
}
\examples{
con<-dbconnect2()
dbListTable(conn = con)
dbDisconnect(con)
}
