% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_ras}
\alias{density_ras}
\title{Surface modeling of kernel density.}
\usage{
density_ras(x, resto = 500, extent.ras = NULL, proj = "+init=epsg:3857",
  negative.keep = F, adjust = 1, ...)
}
\arguments{
\item{x}{a SpatialPointsDataframe object}

\item{resto}{numeric.resolution of output raster}

\item{extent.ras}{extent object.spaital extent of output raster.If missing,the extent of x will be used.}

\item{negative.keep}{logical value indicating whether to keep negative value.If false, all negative value will be assigned to zero.}

\item{adjust}{numerical value indicating adjust value multiplied to bandwidth.The default value is 1.}
}
\value{
RasterLayer
}
\description{
This function calculate the kernel density raster according to a SpatialPointDataframe object.
}
\examples{
n<-100
x<-rnorm(n,mean=100)
y<-rnorm(n,mean=40)
xy<-cbind(x,y)
vals<-1:n
p<-data.frame(xy,name=vals)
coordinates(p)<-~x+y
proj4string(p)<-'+init=epsg:4326'
ras.dens<-density_ras(p)
plot(ras.dens)
}
