% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_shannon.R
\name{diversity_shannon}
\alias{diversity_shannon}
\title{Surface modeling of shannon diversity index.}
\usage{
diversity_shannon(x, poi_type = NULL, res = 500, extent.ras = NULL,
  proj = "+init=epsg:3857")
}
\arguments{
\item{x}{a SpatialPointsDataframe object}

\item{poi_type}{character.column name indicating POI types}

\item{extent.ras}{extent object.spaital extent of output raster.If missing,the extent of x will be used.}

\item{proj}{character.the project string of output.The defalut value is epsg:3857}
}
\value{
RasterLayer
}
\description{
This function calculate shannon diversity index raster according to a SpatialPointDataframe object(POI).
}
