% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_etl.R
\name{extend.extent}
\alias{extend.extent}
\title{extentd the extent by the distance(unit meters)}
\usage{
extend.extent(extent.city = NULL, distance = 50000)
}
\arguments{
\item{extent.city}{extent object}

\item{distance}{a numerica value which indicates the distance will be extended to the extent of a city.}
}
\value{
a extended extent object
}
\description{
function which extentd a specified extent object by the distance.
}
\details{
The unit of distance is meters.
}
