% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getDensity}
\alias{getDensity}
\title{compute density raster for a POI.}
\usage{
getDensity(obj_name, list = NULL, ...)
}
\arguments{
\item{obj_name}{character. name of POI SpatialPointDataframe object}

\item{list}{list object.If not NULL, variable will be extracted from list by the name of obj_name.}

\item{...}{other parameters to be passed to \code{\link{density_ras}}}
}
\value{
RasterLayer
}
\description{
function which compute density raster for a POI specified by a string.
}
\seealso{
\code{\link{density_ras}}
}
