% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha_poi_mindist.R
\name{getMindist}
\alias{getMindist}
\title{function which compute minimum distance raster between specified POI and ha.}
\usage{
getMindist(poi_name, ha = ha_poi.sp, list = NULL)
}
\arguments{
\item{poi_name}{character. name of POI SpatialPointDataframe object}

\item{list}{list object.If not NULL, variable will be extracted from list by the name of obj_name.}

\item{ext}{extent of output raster}
}
\value{
dataframe
}
\description{
function which compute minimum distance raster between specified POI and ha.
}
