% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_etl.R
\name{getpoi.ras}
\alias{getpoi.ras}
\title{generating of POI rasters}
\usage{
getpoi.ras(month_offset = -2, resto = 500, con = NULL, tabln.vec = NULL)
}
\arguments{
\item{month_offset}{offset of month for loading data.The default is -2 indiciting reading data in the previous moth before the last month}

\item{resto}{interger value indicing resolution of raster}

\item{con}{connection of db}
}
\value{
list of POI rasters
}
\description{
This function will computing distances between ha and POI, density ,and diversity raster of POI
}
