% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_etl.R
\name{getpoi.ras2}
\alias{getpoi.ras2}
\title{generating of POI rasters(ver.2)}
\usage{
getpoi.ras2(yearmonth = paste(as.numeric(format(seq(Sys.Date(), by =
  "-2 month", length = 2)[2], "\%Y")), as.numeric(format(seq(Sys.Date(), by =
  "-2 month", length = 2)[2], "\%m")), sep = "-"), resto = 500, con = NULL,
  tabln.vec = NULL)
}
\arguments{
\item{yearmonth}{character values indicting year and month. The default is previous month before the last month.}

\item{resto}{interger value indicing resolution of raster}

\item{con}{connection of db}
}
\value{
list of POI data
}
\description{
This function will computing distances between ha and POI, density ,and diversity raster of POI
}
