% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{install.dependencies}
\alias{install.dependencies}
\title{setting up dependencies}
\usage{
install.dependencies()
}
\description{
This function will install any dependencies of this package
}
\examples{
install.dependencies()
}
