% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{killDbConnections}
\alias{killDbConnections}
\title{kill all db connections}
\usage{
killDbConnections()
}
\description{
This function is useful for cleanning up any MySQL connections.
}
\examples{
killDbConnections()
}
