% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl.R
\name{loadData}
\alias{loadData}
\title{loading data}
\usage{
loadData(proptype = "11", con = dbconnect(), month_offset = -1,
  remove.dup = T)
}
\arguments{
\item{proptype}{type code of property.The default is 11 indicing housing}

\item{con}{connection od db server}

\item{month_offset}{offset value of month.The defalut is -1 which indicing reading data in the previous month}

\item{remove.dup}{logic parameter indicting wherther to delete duplicated coordinates,The defalut is T}
}
\value{
basic data list
}
\description{
Loding data for gwr.avm
}
