% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl.R
\name{loadData2}
\alias{loadData2}
\title{loading data(ver.2)}
\usage{
loadData2(con = dbconnect(), proptype = "11",
  yearmonth = paste(as.numeric(format(seq(Sys.Date(), by = "-2 month", length
  = 2)[2], "\%Y")), as.numeric(format(seq(Sys.Date(), by = "-2 month", length =
  2)[2], "\%m")), sep = "-"), remove.dup = T)
}
\arguments{
\item{con}{connection od db server}

\item{proptype}{type code of property.The default is 11 indicing housing}

\item{yearmonth}{character values indicting year and month. The default is previous month before the last month.}

\item{remove.dup}{logic parameter indicting wherther to delete duplicated coordinates,The defalut is T}
}
\value{
basic data
}
\description{
Loding data for gwr.avm
}
