% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.store.path}
\alias{model.store.path}
\title{the path to store the models}
\usage{
model.store.path(avm.model.store.path = "./models.avm")
}
\arguments{
\item{avm.model.store.path}{The path of saving models. Default is './models.avm'}
}
\description{
This function can set the path of storing models.
}
