% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl.R
\name{outofChina}
\alias{outofChina}
\title{Out-of-China tesing.}
\usage{
outofChina(lat, lon)
}
\arguments{
\item{lat}{lat,required}

\item{lon}{lon,required}
}
\value{
FALSE indiciting not out of china
}
\description{
THis function will be useful for cleanning data out of china
}
\examples{
outofChina(lat = 120,lon =30 )
outofChina(lon = 120,lat = 30 )
}
