% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppsale.R
\name{ppsale}
\alias{ppsale}
\title{joining ha and price data}
\usage{
ppsale(tabln.vec = NULL, proptype = "11", modelpath1 = NULL,
  city_code = NULL)
}
\arguments{
\item{tabln.vec}{basic data provided by \code{\link{loadData}}}

\item{proptype}{type code of properties.The defalut is 11 indiciting housing}
}
\value{
data including ha info and ha price
}
\description{
This function can join price and ha info
}
