% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl2.R
\name{read.deal}
\alias{read.deal}
\title{read deal data from database}
\usage{
read.deal(con = NULL, table.names = "qd_forsale",
  year = as.numeric(format(seq(Sys.Date(), by = "-2 month", length = 2)[2],
  "\%Y")), month = as.numeric(format(seq(Sys.Date(), by = "-2 month", length =
  2)[2], "\%m")), proptype = "11", minsale = 2000, minrent = 1,
  var.except = NULL)
}
\arguments{
\item{con}{connection of db server,which returned by \code{\link{dbconnect}} or similar function in the RMySQL}

\item{table.names}{table names of dead data. The defalut is 'qd_forsale'}

\item{year}{year,the default value is the year of date with the previous month of the last month}

\item{month}{month,,the default value is the month of date with the previous month of the last month}

\item{proptype}{type code of property.The default is 11}

\item{minsale}{Allowed minimum value of saleprice.The defalut is 2000}

\item{minrent}{Allowed minimum value of rentprice.The default is 1}
}
\value{
list of AVM models
}
\description{
This function read deal data according given parameters
}
