% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl2.R
\name{read.sql}
\alias{read.sql}
\title{read data by a sql statement}
\usage{
read.sql(con = NULL, sql_cmd = NULL)
}
\arguments{
\item{con}{connection of db server,which returned by \code{\link{dbconnect}} or similar function in the RMySQL}

\item{sql_cmd}{sql query statement}
}
\value{
model list of avm
}
\description{
This function will read data by given sql query statement
}
