% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl2.R
\name{remove.spatial.outlier}
\alias{remove.spatial.outlier}
\title{remove spatial outliers}
\usage{
remove.spatial.outlier(object, city_code = NULL, x = "x", y = "y",
  probs = 0.9995, type = 8, xtrim = 0.2, ytrim = 0.2)
}
\arguments{
\item{object}{data object, can be dataframe or spaital dataframe.}

\item{city_code}{the code of city to find boundary of city,required .}

\item{x}{field name of lon,the default is 'x'}

\item{y}{field name of lat,the default is 'y'}

\item{probs}{prob of quantile for detecting abnormal coordiantes}

\item{type}{type of quantile, see type in \code{\link{quantile}}}

\item{xtrim}{trim for x in quantile}

\item{ytrim}{trim for y in quantile}
}
\value{
data cleaned
}
\description{
This function help to drop sptaital outliers
}
