% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avm.R
\name{select_model}
\alias{select_model}
\title{This function is used for choosing the best model by the indicator of RMSE}
\usage{
select_model(models, testdata, model.list)
}
\arguments{
\item{models}{avmmodels made before}

\item{testdata}{test data from data.sets, data.sets$test}

\item{model.list}{models}
}
\value{
the best model
}
\description{
This function is used for choosing the best model by the indicator of RMSE
}
