% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl.R
\name{sp.pts}
\alias{sp.pts}
\title{vectorize points.}
\usage{
sp.pts(data_ = NULL, con = dbconnect(), x = "x", y = "y",
  sqlcmd = NULL, remove.dup = T)
}
\arguments{
\item{data_}{raw data. When \code{data_} is not null, \code{con} will be ignored}

\item{con}{connection of db. Either \code{data_} or \code{con} must not be null.}

\item{x}{fieldname of lon.The default is 'x'}

\item{y}{fieldname of lat,The default is 'y'}

\item{sqlcmd}{sql query statement}

\item{remove.dup}{logic parameter indicting wherther to delete duplicated coordinates,The defalut is T}
}
\value{
SpatialPointDataFrame
}
\description{
This function is useful for spatialize raw table data with coordinates
}
\examples{
plot(sp.pts(x='x',y='y'))
}
