% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetl.R
\name{which.in.china}
\alias{which.in.china}
\title{finding which rows of coordinates are inside China.}
\usage{
which.in.china(x)
}
\arguments{
\item{x}{SpatialPointDataFrame}
}
\value{
index of which data are inside China.
}
\description{
This function is for finding which rows of coordinates are inside China.
}
\examples{
data(meuse.grid)
coordinates(meuse.grid) <- ~x+y
which.in.china(meuse.grid)
}
